backend/requirements.txt
fastapi
uvicorn
python-multipart
firebase-admin
supabase
python-dotenv
librosa
soundfile
parselmouth
scikit-learn
joblib
numpy
reportlab
